pruebas

+++ src/main/java/edu/eci/cvds/sampleprj/dao/mybatis/mappers/ClienteMapper.java
@@ -8,7 +8,7 @@ import edu.eci.cvds.samples.entities.Cliente;
 import edu.eci.cvds.samples.entities.ItemRentado;
 public interface ClienteMapper {
 
-    public Cliente consultarCliente(@Param("idcli") int id);
+    public Cliente consultarCliente(@Param("idcli") long id);
 
     /**
      * Registrar un nuevo item rentado asociado al cliente identificado
@@ -19,7 +19,7 @@ public interface ClienteMapper {
      * @param fechafin
      */
 
-    public void agregarItemRentadoACliente(@Param("idcli") int id,
+    public void agregarItemRentadoACliente(@Param("idcli") long id,
                                            @Param("idit")int idit,
                                            @Param("fechainicio")Date fechainicio,
                                            @Param("fechafin")Date fechafin);
